[MOD] 한꺼번에 input창으로 변하는 문제 해결
